requirements.txt
aiohappyeyeballs==2.4.8
aiohttp==3.11.13
aiohttp-jinja2==1.6
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
attrs==25.1.0
certifi==2025.1.31
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.11
fastapi-cli==0.0.7
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.5
linkify-it-py==2.0.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdit-py-plugins==0.4.2
mdurl==0.1.2
msgpack==1.1.0
multidict==6.1.0
platformdirs==4.3.6
propcache==0.3.0
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
rich==13.9.4
rich-toolkit==0.13.2
shellingham==1.5.4
sniffio==1.3.1
starlette==0.46.1
textual==2.1.2
textual-dev==1.7.0
textual-serve==1.1.1
tree-sitter==0.24.0
tree-sitter-bash==0.23.3
tree-sitter-css==0.23.2
tree-sitter-go==0.23.4
tree-sitter-html==0.23.2
tree-sitter-java==0.23.5
tree-sitter-javascript==0.23.1
tree-sitter-json==0.24.8
tree-sitter-markdown==0.3.2
tree-sitter-python==0.23.6
tree-sitter-regex==0.24.3
tree-sitter-rust==0.23.2
tree-sitter-sql==0.3.7
tree-sitter-toml==0.7.0
tree-sitter-xml==0.7.0
tree-sitter-yaml==0.7.0
typer==0.15.2
typing_extensions==4.12.2
uc-micro-py==1.0.3
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0.1
yarl==1.18.3
